Add offlineimap configuration and script. Mutt not yet using it.
The offlineimaprc file is a basic setup to sync my Gmail account. It
uses an external python script, offlineimap.py to do the folder name
translations between the local and remote repositories.
That script also has a function to retrieve a password from OSX's
keychain so that offlineimap can be run on a regular basis without
requesting password or having a password in a file.